+++ crates/rethnet_eth/src/remote/methods/hardhat/add_compilation_result.rs
